subscription_service/requirements.txt
flask
mysql-connector-python
requests
flask-swagger-ui
